+++ lib/atoms/text/base/zwapText.dart
@@ -106,9 +106,8 @@ Size getTextSize(String text, ZwapTextType textType, {double? maxWidth}) {
 }
 
 Size getTextSizeFromCustomStyle(String text, TextStyle textStyle, {double? maxWidth}) {
-  final TextPainter textPainter =
-      TextPainter(text: TextSpan(text: text, style: textStyle), maxLines: 1, textDirection: TextDirection.ltr)
-        ..layout(minWidth: 0, maxWidth: maxWidth ?? double.infinity);
+  final TextPainter textPainter = TextPainter(text: TextSpan(text: text, style: textStyle), maxLines: 1, textDirection: TextDirection.ltr)
+    ..layout(minWidth: 0, maxWidth: maxWidth ?? double.infinity);
   return textPainter.size;
 }